### 2018-08-05 #875 @whedon
[PRE REVIEW]: Containershare: Open Source Registry to build, test, deploy with CircleCI
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://github.com/vsoch/containershare" target ="_blank">https://github.com/vsoch/containershare</a>
**Version:** v.0.0.1
