### 2018-08-06 #875 @brainstorm
Happy to review this one, @arfon ;)
