### 2018-08-06 #878 @whedon
[REVIEW]: Containershare: Open Source Registry to build, test, deploy with CircleCI
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** <a href="https://www.github.com/vsoch/containershare" target ="_blank">https://www.github.com/vsoch/containershare</a>
**Version:** 0.0.3

- [x] **Version:** Does the release version given match the GitHub release (v.0.0.1)?
