### 2018-08-06 #878 @brainstorm
Alright, let's start! Thanks @arfon 👍

@vsoch Your project/repo seems quite well put together, but I have one question you'll have to motivate a bit more on `paper.md`: How does containershare compare/supersede/complement/replace/deprecate http://biocontainers.pro/  (https://github.com/BioContainers/containers) and other similar initiatives in this space from i.e ELIXIR-like initiatives in this space?

In other words, `paper.bib` needs a bit more refs to motivate the "Statement of need", IMHO ;)
