### 2018-08-06 #878 @brainstorm
@vsoch: Minor wrinkle: I would adhere to [semver](https://semver.org/) better by dropping the initial `v` in your releases, i.e `0.0.1` instead of `v0.0.1`.
