### 2018-08-06 #878 @brainstorm
Could you also have a small mention on how to run the unit tests? ;)
