### 2018-08-06 72dfdcb
first shot at adding biocontainers reference
