+++ paper/paper.bib
@@ -5,6 +5,42 @@
 }
 
 
+@ARTICLE{Da_Veiga_Leprevost2017-xx,
+  title    = "{BioContainers}: an open-source and community-driven framework
+              for software standardization",
+  author   = "da Veiga Leprevost, Felipe and Gr{\"u}ning, Bj{\"o}rn A and Alves
+              Aflitos, Saulo and R{\"o}st, Hannes L and Uszkoreit, Julian and
+              Barsnes, Harald and Vaudel, Marc and Moreno, Pablo and Gatto,
+              Laurent and Weber, Jonas and Bai, Mingze and Jimenez, Rafael C
+              and Sachsenberg, Timo and Pfeuffer, Julianus and Vera Alvarez,
+              Roberto and Griss, Johannes and Nesvizhskii, Alexey I and
+              Perez-Riverol, Yasset",
+  abstract = "Motivation: BioContainers (biocontainers.pro) is an open-source
+              and community-driven framework which provides platform
+              independent executable environments for bioinformatics software.
+              BioContainers allows labs of all sizes to easily install
+              bioinformatics software, maintain multiple versions of the same
+              software and combine tools into powerful analysis pipelines.
+              BioContainers is based on popular open-source projects Docker and
+              rkt frameworks, that allow software to be installed and executed
+              under an isolated and controlled environment. Also, it provides
+              infrastructure and basic guidelines to create, manage and
+              distribute bioinformatics containers with a special focus on
+              omics technologies. These containers can be integrated into more
+              comprehensive bioinformatics pipelines and different
+              architectures (local desktop, cloud environments or HPC
+              clusters). Availability and Implementation: The software is
+              freely available at github.com/BioContainers/. Contact:
+              yperez@ebi.ac.uk.",
+  journal  = "Bioinformatics",
+  volume   =  33,
+  number   =  16,
+  pages    = "2580--2582",
+  month    =  aug,
+  year     =  2017,
+  language = "en"
+}
+
 @MISC{noauthor_undated-vt,
   title       = "{GitHub} Pages",
   abstract    = "Websites for you and your projects, hosted directly from your
