### 2018-08-06 d09b7a1
tweaks to README to add testing instructions
