+++ .circleci/config.yml
@@ -26,7 +26,7 @@ jobs:
           command: |
             wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
             /bin/bash Miniconda3-latest-Linux-x86_64.sh -b 
-            $HOME/miniconda3/bin/python -m pip install pyaml containershare==0.0.12
+            $HOME/miniconda3/bin/python -m pip install pyaml containershare==0.0.14
             cd $HOME/containershare/docs
             $HOME/miniconda3/bin/python ../tests/circle_urls.py $HOME/containershare/docs/_site
       - run:
