### 2018-08-06 #878 @vsoch
Thanks for the speedy update @brainstorm and @arfon (and @whedon!) I've done the following changes to address the review comments:

 - a [first shot](https://github.com/vsoch/containershare#local-tests) at explaining the difference between containershare and biocontainers. I'm not terribly familiar with biocontainers, but it seems to serve as a general listing of bioinformatics containers, and isn't intended to be a general template to extend to other purposes and provide extended metadata (beyond an image manifest). It doesn't take a distributed strategy of requiring each container repository to serve not only different kinds of metadata, but also a LICENSE and README, and it seems to be less lightweight in requiring a registry server for the API. Containershare serves it's api from Github pages without additional dependencies. To summarize, I'd say containershare is a more lightweight, customizable solution that is picker about its containers spilling their tiny souls for inspection.
 - The tag will be changed to remove the v on the next tagged release (after our changes and discussion!) to adhere to semvar.
 - I had not provided clear steps for running tests locally, and actually since on CircleCI the tests are only run for newly added files, I had missed a test not passing (the julia-share was missing a license!). I've updated both the containershare testing software (version to 0.0.14), the [missing license](https://github.com/vsoch/julia-share/pull/1/files) along with adding instructions to the README to test locally. I'm really glad you asked about this!

I think that should be good for this first tweak - take a look and let me know what needs more work! I'll also note that I have several new container templates in the queue - the exciting thing about this work is that it isn't a static "it's done now and forever" thing, it's going to be extended to have lots of different container types (templates) that a user can add to the containershare!
