### 2018-08-06 #878 @vsoch
I just edited @whedon 's version up to what we will tag as the final release, after review:

```bash
**Version:** 0.0.3
```
I didn't realize I could do that :)
