### 2018-08-06 #878 @vsoch
heyo everyone! For a full circle test of the above changes, I added a repo2docker-julia template just now, and everything worked as I would want and expect! You can see:

 - [the CircleCI testing here](https://circleci.com/gh/vsoch/containershare/26?utm_campaign=vcs-integration-link&utm_medium=referral&utm_source=github-build-link)
 - [the pull request here](https://github.com/vsoch/containershare/pull/11)

And upon merge, the [table](https://vsoch.github.io/containershare/) and associated linked metadata was updated.
