### 2018-08-06 #878 @brainstorm
Cool, thanks for the quick changes @vsoch, all LGTM now.

I would suggest to add a bib reference to biocontainers since you mention them... not sure they even have a preprint though, care to contact and ask them?: http://biocontainers.pro/#contact

Other than that, all clear, good job! ;)
