### 2018-08-07 #878 @vsoch
oh hmm, I definitely did that! I think Mr.Whedon just needs to rebuild the pdf? Let me see if I can figure that out... @whedon commands
