### 2018-08-07 #878 @vsoch
All set! version 0.0.3 with the final work is here --> https://zenodo.org/account/settings/github/repository/vsoch/containershare# pointing to https://github.com/vsoch/containershare/releases/tag/0.0.3.

The link for the doi provided is https://zenodo.org/badge/latestdoi/142066803 (although not found yet?)

Let me know if you need anything else!
